Refactor: move source into subdirectory
